Extract callwithchat test app
