+++ packages/react-composites/tests/browser/callwithchat/fixture.ts
@@ -9,7 +9,7 @@ import { test as base } from '@playwright/test';
 import path from 'path';
 
 const SERVER_URL = 'http://localhost';
-const APP_DIR = path.join(__dirname, 'app');
+const APP_DIR = path.join(__dirname, '../../../app/callwithchat');
 
 type CallWithChatWorkerFixture = WorkerFixture<CallWithChatUserType>;
